Done tugas 8

+++ EL3111_2_20221006_13220034/1_Lab/Tugas_8/coba_double.c
@@ -11,10 +11,10 @@
 
 void coba(int* x, int* y, int* z){
     // Kamus
-    int a;
-    int b;
-    int c;
-    int d;
+    double a;
+    double b;
+    double c;
+    double d;
     // Algoritma
     a = *x;
     b = *y;
